# products/cortex/apps/web/openapi.json
{
  "components": {
    "schemas": {
...
      "AnswerResponse": {
        "properties": {
          "result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AnswerResult"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": ["result"],
        "title": "AnswerResponse",
        "type": "object"
      },
...
      "AppendTurnRequest": {
        "additionalProperties": false,
        "properties": {
          "content": {
            "maxLength": 8000,
            "minLength": 1,
            "title": "Content",
            "type": "string"
          }
        },
        "required": ["content"],
        "title": "AppendTurnRequest",
        "type": "object"
      },
      "AppendTurnResponse": {
        "properties": {
          "conversation": {
            "$ref": "#/components/schemas/ConversationResponse"
          },
          "model": {
            "title": "Model",
            "type": "string"
          },
          "usage": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TokenUsage"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": ["conversation", "model"],
        "title": "AppendTurnResponse",
        "type": "object"
      },
...
      "ChatMessage": {
        "additionalProperties": false,
        "properties": {
          "content": {
            "maxLength": 8000,
            "minLength": 1,
            "title": "Content",
            "type": "string"
          },
          "role": {
            "enum": ["system", "user", "assistant"],
            "title": "Role",
            "type": "string"
          }
        },
        "required": ["role", "content"],
        "title": "ChatMessage",
        "type": "object"
      },
      "ChatTurnRequest": {
        "additionalProperties": false,
        "properties": {
          "messages": {
            "items": {
              "$ref": "#/components/schemas/ChatMessage"
            },
            "maxItems": 50,
            "minItems": 1,
            "title": "Messages",
            "type": "array"
          }
        },
        "required": ["messages"],
        "title": "ChatTurnRequest",
        "type": "object"
      },
      "ChatTurnResponse": {
        "additionalProperties": false,
        "properties": {
          "message": {
            "$ref": "#/components/schemas/ChatMessage"
          },
          "model": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Model",
            "type": "string"
          },
          "usage": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TokenUsage"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": ["message", "model"],
        "title": "ChatTurnResponse",
        "type": "object"
      },
      "ConversationListResponse": {
        "properties": {
          "conversations": {
            "items": {
              "$ref": "#/components/schemas/ConversationSummary"
            },
            "title": "Conversations",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          }
        },
        "required": ["conversations", "limit", "offset"],
        "title": "ConversationListResponse",
        "type": "object"
      },
...
      "ConversationResponse": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "messages": {
            "items": {
              "$ref": "#/components/schemas/ConversationMessageResponse"
            },
            "title": "Messages",
            "type": "array"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": ["id", "title", "created_at", "updated_at", "messages"],
        "title": "ConversationResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LookupRequest": {
        "additionalProperties": false,
        "properties": {
          "query": {
            "maxLength": 500,
            "minLength": 1,
            "title": "Query",
            "type": "string"
          }
        },
        "required": ["query"],
        "title": "LookupRequest",
        "type": "object"
      },
      "LookupResponse": {
        "properties": {
          "result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LookupResult"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": ["result"],
        "title": "LookupResponse",
        "type": "object"
      },
...
      "TokenUsage": {
        "additionalProperties": false,
        "properties": {
          "input_tokens": {
            "minimum": 0.0,
            "title": "Input Tokens",
            "type": "integer"
          },
          "output_tokens": {
            "minimum": 0.0,
            "title": "Output Tokens",
            "type": "integer"
          }
        },
        "required": ["input_tokens", "output_tokens"],
        "title": "TokenUsage",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Cortex",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/chat/turn": {
      "post": {
        "operationId": "chat_turn_chat_turn_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatTurnRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatTurnResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Chat Turn",
        "tags": ["chat"]
      }
    },
    "/conversations": {
      "get": {
        "operationId": "list_conversations_conversations_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "maximum": 10000,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConversationListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Conversations",
        "tags": ["conversations"]
      },
      "post": {
        "operationId": "create_conversation_conversations_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConversationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Conversation",
        "tags": ["conversations"]
      }
    },
    "/conversations/{conversation_id}": {
      "get": {
        "operationId": "get_conversation_conversations__conversation_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "conversation_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Conversation Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConversationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Conversation",
        "tags": ["conversations"]
      }
    },
    "/conversations/{conversation_id}/turns": {
      "post": {
        "operationId": "append_conversation_turn_conversations__conversation_id__turns_post",
        "parameters": [
          {
            "in": "path",
            "name": "conversation_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Conversation Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppendTurnRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppendTurnResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Append Conversation Turn",
        "tags": ["conversations"]
      }
    },
    "/conversations/{conversation_id}/turns/stream": {
      "post": {
        "operationId": "stream_conversation_turn_conversations__conversation_id__turns_stream_post",
        "parameters": [
          {
            "in": "path",
            "name": "conversation_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Conversation Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppendTurnRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Stream Conversation Turn",
        "tags": ["conversations"]
      }
    },
...
    "/knowledge/answer": {
      "post": {
        "operationId": "answer_knowledge_knowledge_answer_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LookupRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnswerResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Answer Knowledge",
        "tags": ["knowledge"]
      }
    },
    "/knowledge/lookup": {
      "post": {
        "operationId": "lookup_knowledge_knowledge_lookup_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LookupRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LookupResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Lookup Knowledge",
        "tags": ["knowledge"]
      }
    }
  }
}